scaffolded out the UI -- data flows included, will update query once id is nullable
