+++ apps/protocol-frontend/src/components/ContributionsByDateShell.tsx
@@ -1,107 +1,41 @@
-import React from 'react';
-import { Flex, Box, Text, useBreakpointValue } from '@chakra-ui/react';
-import { ResponsiveTimeRange } from '@nivo/calendar';
-import { GovrnTheme } from '@govrn/protocol-ui';
-import { subWeeks } from 'date-fns';
+import useContributionCountInRange from '../hooks/useContributionCount';
+import ContributionsByDateChart from './ContributionsByDateChart';
+import { TODAY_DATE, YEAR } from '../utils/constants';
+import { subWeeks, startOfDay, endOfDay } from 'date-fns';
 
-type ContributionCount = {
-  date: string;
-  count: number;
-};
-
-interface ContributionsByDateShellProps {
-  contributionsCount: ContributionCount[];
-  startDateOffset?: number;
+interface ContributionByDateShellProps {
+  guildIds: number[];
+  // startDate: Date;
+  // endDate: Date;
 }
 
-const brandColors = GovrnTheme.colors.brand.primary;
-
-const brandColorMap = [
-  '#ffb4e2',
-  '#fb84ce',
-  '#f854ba',
-  '#f526a6',
-  '#db0f8d',
-  '#ab076d',
-  '#76024e',
-];
-
-const ContributionsByDateShell = ({
-  contributionsCount,
-  startDateOffset,
-}: ContributionsByDateShellProps) => {
-  const isMobile = useBreakpointValue({ base: true, lg: false });
-  const contributionsCountMap = contributionsCount.map(contribution => {
-    return {
-      day: contribution.date,
-      value: contribution.count,
-    };
+const ContributionByDateShell = ({
+  guildIds,
+}: // startDate,
+// endDate,
+ContributionByDateShellProps) => {
+  const {
+    isFetching,
+    isLoading,
+    isError,
+    data: contributionsByDate,
+  } = useContributionCountInRange({
+    startDate: subWeeks(startOfDay(TODAY_DATE), YEAR),
+    endDate: endOfDay(TODAY_DATE),
+    // startDate: startDate,
+    // endDate: endDate,
+    guildIds,
+    excludeUnassigned: true,
   });
 
   return (
-    <Flex direction="column" paddingBottom={4} paddingX={{ base: 4, lg: 0 }}>
-      <Flex
-        direction="column"
-        minHeight={{ base: '5rem', lg: '10rem' }}
-        height={{ base: '5rem', lg: '10rem' }}
-        width="100%"
-        maxWidth="100vw"
-      >
-        {contributionsCountMap.length !== 0 && (
-          <ResponsiveTimeRange
-            data={contributionsCountMap.filter(
-              contribution => contribution.value !== 0,
-            )}
-            from={
-              startDateOffset
-                ? subWeeks(new Date(), startDateOffset)
-                : subWeeks(new Date(), 52)
-            }
-            to={new Date()}
-            weekdayTicks={isMobile ? [] : [1, 3, 5]}
-            emptyColor="#eeeeee"
-            colors={[
-              brandColors[100],
-              brandColors[200],
-              brandColors[300],
-              brandColors[400],
-              brandColors[500],
-              brandColors[600],
-              brandColors[700],
-            ]}
-            dayRadius={1}
-            margin={
-              isMobile
-                ? { top: 20, right: 10, bottom: 0, left: 10 }
-                : { top: 20, right: 20, bottom: 20, left: 20 }
-            }
-            dayBorderWidth={2}
-            dayBorderColor="#ffffff"
-          />
-        )}
-      </Flex>
-      <Flex
-        direction="row"
-        alignItems="center"
-        justifyContent={{ base: 'flex-start', lg: 'center' }}
-      >
-        <Text as="span" fontSize="sm" fontWeight="normal" paddingRight={1}>
-          Less
-        </Text>
-        {brandColorMap.map(color => (
-          <Box
-            key={color}
-            backgroundColor={color}
-            width="15.36px"
-            height="15.36px"
-          />
-        ))}
-        <Text as="span" fontSize="sm" fontWeight="normal" paddingLeft={1}>
-          More
-        </Text>
-      </Flex>
-    </Flex>
+    <ContributionsByDateChart
+      contributionsCount={contributionsByDate}
+      isFetching={isFetching}
+      isLoading={isLoading}
+      isError={isError}
+    />
   );
 };
 
-export default ContributionsByDateShell;
+export default ContributionByDateShell;
